# spring/src/main/java/com/example/meetingrooms/api/BookingsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0-SNAPSHOT).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Booking;
import org.openapitools.model.BookingRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-26T17:15:28.044987462Z[Etc/UTC]", comments = "Generator version: 7.21.0-SNAPSHOT")
@Validated
@Tag(name = "bookings", description = "the bookings API")
public interface BookingsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_BOOKINGS_BOOKING_ID_DELETE = "/bookings/{booking_id}";
    /**
     * DELETE /bookings/{booking_id}
     * Отмена существующего бронирования
     *
     * @param bookingId ID бронирования (required)
     * @return Бронирование успешно отменено (status code 204)
     *         or Бронирование не найдено (status code 404)
     */
    @Operation(
        operationId = "bookingsBookingIdDelete",
        description = "Отмена существующего бронирования",
        responses = {
            @ApiResponse(responseCode = "204", description = "Бронирование успешно отменено"),
            @ApiResponse(responseCode = "404", description = "Бронирование не найдено")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = BookingsApi.PATH_BOOKINGS_BOOKING_ID_DELETE
    )
    default ResponseEntity<Void> bookingsBookingIdDelete(
        @NotNull @Min(value = 1) @Parameter(name = "booking_id", description = "ID бронирования", required = true, in = ParameterIn.PATH) @PathVariable("booking_id") Integer bookingId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_BOOKINGS_POST = "/bookings";
    /**
     * POST /bookings
     * Создание нового бронирования переговорки
     *
     * @param bookingRequest  (required)
     * @return Бронирование успешно создано (status code 201)
     *         or Ошибка в данных бронирования (status code 400)
     *         or Время уже занято (status code 409)
     *         or Переговорка не найдена (status code 404)
     */
    @Operation(
        operationId = "bookingsPost",
        description = "Создание нового бронирования переговорки",
        responses = {
            @ApiResponse(responseCode = "201", description = "Бронирование успешно создано", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Booking.class))
            }),
            @ApiResponse(responseCode = "400", description = "Ошибка в данных бронирования"),
            @ApiResponse(responseCode = "409", description = "Время уже занято"),
            @ApiResponse(responseCode = "404", description = "Переговорка не найдена")
        },
        security = {
            @SecurityRequirement(name = "cookieAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = BookingsApi.PATH_BOOKINGS_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Booking> bookingsPost(
        @Parameter(name = "BookingRequest", description = "", required = true) @Valid @RequestBody BookingRequest bookingRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"room_id\" : 1, \"room_name\" : \"Конференц-зал А\", \"start_time\" : \"2024-01-20T10:00:00Z\", \"booked_by\" : \"alex@company.com\", \"end_time\" : \"2024-01-20T11:30:00Z\", \"created_at\" : \"2024-01-19T15:30:00Z\", \"id\" : 42, \"title\" : \"Спринт-ревью\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_BOOKING = "/bookings/{booking_id}";
    /**
     * GET /bookings/{booking_id}
     * Получение детальной информации о конкретном бронировании
     *
     * @param bookingId ID бронирования (required)
     * @return Информация о бронировании (status code 200)
     *         or Бронирование не найдено (status code 404)
     */
    @Operation(
        operationId = "getBooking",
        description = "Получение детальной информации о конкретном бронировании",
        responses = {
            @ApiResponse(responseCode = "200", description = "Информация о бронировании", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Booking.class))
            }),
            @ApiResponse(responseCode = "404", description = "Бронирование не найдено")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BookingsApi.PATH_GET_BOOKING,
        produces = { "application/json" }
    )
    default ResponseEntity<Booking> getBooking(
        @NotNull @Min(value = 1) @Parameter(name = "booking_id", description = "ID бронирования", required = true, in = ParameterIn.PATH) @PathVariable("booking_id") Integer bookingId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"room_id\" : 1, \"room_name\" : \"Конференц-зал А\", \"start_time\" : \"2024-01-20T10:00:00Z\", \"booked_by\" : \"alex@company.com\", \"end_time\" : \"2024-01-20T11:30:00Z\", \"created_at\" : \"2024-01-19T15:30:00Z\", \"id\" : 42, \"title\" : \"Спринт-ревью\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}