# spring/src/main/java/com/example/meetingrooms/api/RoomsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0-SNAPSHOT).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Booking;
import org.springframework.format.annotation.DateTimeFormat;
import java.time.LocalDate;
import org.openapitools.model.Room;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-26T17:15:28.044987462Z[Etc/UTC]", comments = "Generator version: 7.21.0-SNAPSHOT")
@Validated
@Tag(name = "rooms", description = "the rooms API")
public interface RoomsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_ROOMS_GET = "/rooms";
    /**
     * GET /rooms
     * Получение списка всех доступных переговорных комнат
     *
     * @return Список переговорок (status code 200)
     *         or Внутренняя ошибка сервера (status code 500)
     */
    @Operation(
        operationId = "roomsGet",
        description = "Получение списка всех доступных переговорных комнат",
        responses = {
            @ApiResponse(responseCode = "200", description = "Список переговорок", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Room.class)))
            }),
            @ApiResponse(responseCode = "500", description = "Внутренняя ошибка сервера")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = RoomsApi.PATH_ROOMS_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Room>> roomsGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"name\" : \"Конференц-зал А\", \"equipment\" : [ \"projector\", \"whiteboard\" ], \"id\" : 1, \"capacity\" : 10 }, { \"name\" : \"Конференц-зал А\", \"equipment\" : [ \"projector\", \"whiteboard\" ], \"id\" : 1, \"capacity\" : 10 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_ROOMS_ROOM_ID_BOOKINGS_GET = "/rooms/{room_id}/bookings";
    /**
     * GET /rooms/{room_id}/bookings
     * Получение списка забронированных слотов для конкретной переговорки на указанную дату
     *
     * @param roomId ID переговорной комнаты (required)
     * @param date Дата в формате YYYY-MM-DD (required)
     * @return Список бронирований на дату (status code 200)
     *         or Переговорка не найдена (status code 404)
     *         or Неверный формат даты (status code 400)
     */
    @Operation(
        operationId = "roomsRoomIdBookingsGet",
        description = "Получение списка забронированных слотов для конкретной переговорки на указанную дату",
        responses = {
            @ApiResponse(responseCode = "200", description = "Список бронирований на дату", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Booking.class)))
            }),
            @ApiResponse(responseCode = "404", description = "Переговорка не найдена"),
            @ApiResponse(responseCode = "400", description = "Неверный формат даты")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = RoomsApi.PATH_ROOMS_ROOM_ID_BOOKINGS_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Booking>> roomsRoomIdBookingsGet(
        @NotNull @Min(value = 1) @Parameter(name = "room_id", description = "ID переговорной комнаты", required = true, in = ParameterIn.PATH) @PathVariable("room_id") Integer roomId,
        @NotNull @Parameter(name = "date", description = "Дата в формате YYYY-MM-DD", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "date", required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate date
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"room_id\" : 1, \"room_name\" : \"Конференц-зал А\", \"start_time\" : \"2024-01-20T10:00:00Z\", \"booked_by\" : \"alex@company.com\", \"end_time\" : \"2024-01-20T11:30:00Z\", \"created_at\" : \"2024-01-19T15:30:00Z\", \"id\" : 42, \"title\" : \"Спринт-ревью\" }, { \"room_id\" : 1, \"room_name\" : \"Конференц-зал А\", \"start_time\" : \"2024-01-20T10:00:00Z\", \"booked_by\" : \"alex@company.com\", \"end_time\" : \"2024-01-20T11:30:00Z\", \"created_at\" : \"2024-01-19T15:30:00Z\", \"id\" : 42, \"title\" : \"Спринт-ревью\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}