# spring/target/classes/openapi.yaml
openapi: 3.0.3
info:
  description: API для бронирования переговорных комнат в офисе
  title: Бронирование переговорных комнат
  version: 0.1.0
servers:
- url: /
paths:
  /rooms:
    get:
      description: Получение списка всех доступных переговорных комнат
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Room"
                type: array
          description: Список переговорок
        "500":
          description: Внутренняя ошибка сервера
      x-accepts:
      - application/json
  /rooms/{room_id}/bookings:
    get:
      description: Получение списка забронированных слотов для конкретной переговорки
        на указанную дату
      parameters:
      - description: ID переговорной комнаты
        explode: false
        in: path
        name: room_id
        required: true
        schema:
          minimum: 1
          type: integer
        style: simple
      - description: Дата в формате YYYY-MM-DD
        explode: true
        in: query
        name: date
        required: true
        schema:
          example: 2024-01-20
          format: date
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Booking"
                type: array
          description: Список бронирований на дату
        "404":
          description: Переговорка не найдена
        "400":
          description: Неверный формат даты
      x-accepts:
      - application/json
  /bookings:
    post:
      description: Создание нового бронирования переговорки
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
          description: Бронирование успешно создано
        "400":
          description: Ошибка в данных бронирования
        "409":
          description: Время уже занято
        "404":
          description: Переговорка не найдена
      security:
      - cookieAuth: []
      x-content-type: application/json
      x-accepts:
      - application/json
  /bookings/{booking_id}:
    delete:
      description: Отмена существующего бронирования
      parameters:
      - description: ID бронирования
        explode: false
        in: path
        name: booking_id
        required: true
        schema:
          minimum: 1
          type: integer
        style: simple
      responses:
        "204":
          description: Бронирование успешно отменено
        "404":
          description: Бронирование не найдено
      x-accepts:
      - application/json
    get:
      description: Получение детальной информации о конкретном бронировании
      operationId: get_booking
      parameters:
      - description: ID бронирования
        explode: false
        in: path
        name: booking_id
        required: true
        schema:
          minimum: 1
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
          description: Информация о бронировании
        "404":
          description: Бронирование не найдено
      x-accepts:
      - application/json
components:
  schemas:
    Room:
      example:
        name: Конференц-зал А
        equipment:
        - projector
        - whiteboard
        id: 1
        capacity: 10
      properties:
        id:
          description: Уникальный идентификатор комнаты
          example: 1
          type: integer
        name:
          description: Название переговорки
          example: Конференц-зал А
          type: string
        capacity:
          description: Вместимость (человек)
          example: 10
          minimum: 1
          type: integer
        equipment:
          description: Оснащение комнаты
          example:
          - projector
          - whiteboard
          items:
            enum:
            - projector
            - tv
            - whiteboard
            - videoconference
            - speaker
            type: string
          type: array
      required:
      - capacity
      - id
      - name
      type: object
    Booking:
      example:
        room_id: 1
        room_name: Конференц-зал А
        start_time: 2024-01-20T10:00:00Z
        booked_by: alex@company.com
        end_time: 2024-01-20T11:30:00Z
        created_at: 2024-01-19T15:30:00Z
        id: 42
        title: Спринт-ревью
      properties:
        id:
          description: Уникальный идентификатор бронирования
          example: 42
          type: integer
        room_id:
          description: ID переговорки
          example: 1
          type: integer
        room_name:
          description: Название переговорки (для удобства)
          example: Конференц-зал А
          type: string
        booked_by:
          description: "Email или имя того, кто забронировал"
          example: alex@company.com
          type: string
        title:
          description: Название встречи
          example: Спринт-ревью
          type: string
        start_time:
          description: Дата и время начала
          example: 2024-01-20T10:00:00Z
          format: date-time
          type: string
        end_time:
          description: Дата и время окончания
          example: 2024-01-20T11:30:00Z
          format: date-time
          type: string
        created_at:
          description: Дата и время создания бронирования
          example: 2024-01-19T15:30:00Z
          format: date-time
          type: string
      required:
      - booked_by
      - end_time
      - id
      - room_id
      - start_time
      type: object
    BookingRequest:
      example:
        room_id: 1
        start_time: 2024-01-20T14:00:00Z
        end_time: 2024-01-20T15:00:00Z
        title: Планирование спринта
      properties:
        room_id:
          description: ID переговорки
          example: 1
          minimum: 1
          type: integer
        title:
          description: Название встречи (опционально)
          example: Планирование спринта
          type: string
        start_time:
          description: Дата и время начала
          example: 2024-01-20T14:00:00Z
          format: date-time
          type: string
        end_time:
          description: Дата и время окончания
          example: 2024-01-20T15:00:00Z
          format: date-time
          type: string
      required:
      - end_time
      - room_id
      - start_time
      type: object
  securitySchemes:
    cookieAuth:
      in: cookie
      name: session
      type: apiKey